requirements.txt
# New official SDK recommended by Google quickstart
google-genai>=1.0.0

python-dotenv==1.0.1
pydantic>=2.11.0,<3.0.0

# MCP (agentify)
mcp[cli]>=1.2.0

# UI
streamlit>=1.32.0

# testing
pytest>=8.0.0